New scripts
To see an interpolation:
1. Run iterate.py to create the support points.
2. Run interpolate.py on the desired point.
compare.py does the same as interpolate.py but compares the result with an independent grmonty run. Plots both and shows the relative error.
convergence.py computes the L1 norm for a point (M, Mdot, theta) over an array of photon numbers.
convergence2.py does the same but compares each photon number with another run with same photon number.